created filter directory | created test and challenge files | updated readme
