+++ javascript/quick-code-challenges/filter/README.md
@@ -1,0 +1,7 @@
+# Filter
+
+Code Challenges utilizing the [JavaScript filter() method](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter).
+
+## Tests
+
+Run your tests from the console: `npm test filter.test.js`.
